2022-01-02 controller, repository changes

+++ src/main/java/me/varnavsky/review_service/model/review/ProductReviewDto.java
@@ -6,15 +6,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
 public class ProductReviewDto {
-    String productId;
-    BigDecimal reviewAverageScore;
-    Integer reviewCount;
+  String productId;
+  Double reviewAverageScore;
+  Long reviewCount;
 }
